fix:  delete state from routes and move to controllers and add renderless component to templates
