+++ ui/app/components/app-breadcrumbs.js
@@ -1,13 +1,3 @@
-import Component from '@ember/component';
-import { inject as service } from '@ember/service';
-import { reads } from '@ember/object/computed';
-import { tagName } from '@ember-decorators/component';
-import classic from 'ember-classic-decorator';
+import Component from '@glimmer/component';
 
-@classic
-@tagName('')
-export default class AppBreadcrumbs extends Component {
-  @service('breadcrumbs') breadcrumbsService;
-
-  @reads('breadcrumbsService.breadcrumbs') breadcrumbs;
-}
+export default class AppBreadcrumbs extends Component {}
