+++ ui/app/controllers/clients/client.js
@@ -1,0 +1,14 @@
+import Controller from '@ember/controller';
+
+export default class ClientsClientController extends Controller {
+  breadcrumbs() {
+    const model = this.model;
+    if (!model) return [];
+    return [
+      {
+        label: model.get('shortId'),
+        args: ['clients.client', model.get('id')],
+      },
+    ];
+  }
+}
