+++ ui/app/controllers/csi/plugins.js
@@ -2,4 +2,11 @@ import Controller from '@ember/controller';
 
 export default class PluginsController extends Controller {
   isForbidden = false;
+
+  breadcrumbs = [
+    {
+      label: 'Storage',
+      args: ['csi.index'],
+    },
+  ];
 }
