+++ ui/app/controllers/csi/volumes.js
@@ -2,4 +2,11 @@ import Controller from '@ember/controller';
 
 export default class VolumesController extends Controller {
   isForbidden = false;
+
+  breadcrumbs = [
+    {
+      label: 'Storage',
+      args: ['csi.index'],
+    },
+  ];
 }
