+++ ui/app/controllers/jobs.js
@@ -1,3 +1,10 @@
 import Controller from '@ember/controller';
 
-export default class JobsController extends Controller {}
+export default class JobsController extends Controller {
+  breadcrumbs = [
+    {
+      label: 'Jobs',
+      args: ['jobs.index'],
+    },
+  ];
+}
