+++ ui/app/controllers/jobs/job.js
@@ -1,4 +1,5 @@
 import Controller from '@ember/controller';
+import { jobCrumbs } from 'nomad-ui/utils/breadcrumb-utils';
 
 export default class JobController extends Controller {
   queryParams = [
@@ -7,4 +8,6 @@ export default class JobController extends Controller {
     },
   ];
   jobNamespace = 'default';
+
+  breadcrumbs = jobCrumbs(this.model);
 }
