+++ ui/app/controllers/jobs/job/dispatch.js
@@ -1,0 +1,10 @@
+import Controller from '@ember/controller';
+
+export default class JobsJobDispatchController extends Controller {
+  breadcrumbs = [
+    {
+      label: 'Dispatch',
+      args: ['jobs.job.dispatch'],
+    },
+  ];
+}
