+++ ui/app/controllers/jobs/job/task-group.js
@@ -5,6 +5,7 @@ import Controller from '@ember/controller';
 import { action, computed, get } from '@ember/object';
 import { scheduleOnce } from '@ember/runloop';
 import intersection from 'lodash.intersection';
+import { qpBuilder } from 'nomad-ui/utils/classes/query-params';
 import Sortable from 'nomad-ui/mixins/sortable';
 import Searchable from 'nomad-ui/mixins/searchable';
 import WithNamespaceResetting from 'nomad-ui/mixins/with-namespace-resetting';
@@ -140,5 +141,20 @@ export default class TaskGroupController extends Controller.extend(
 
   setFacetQueryParam(queryParam, selection) {
     this.set(queryParam, serialize(selection));
+  }
+  get breadcrumbs() {
+    const model = this.model;
+    if (!model) return [];
+    return [
+      {
+        label: model.get('name'),
+        args: [
+          'jobs.job.task-group',
+          model.get('job'),
+          model.get('name'),
+          qpBuilder({ jobNamespace: model.get('job.namespace.name') || 'default' }),
+        ],
+      },
+    ];
   }
 }
