+++ ui/app/controllers/servers.js
@@ -2,4 +2,11 @@ import Controller from '@ember/controller';
 
 export default class ServersController extends Controller {
   isForbidden = false;
+
+  breadcrumbs = [
+    {
+      label: 'Servers',
+      args: ['servers.index'],
+    },
+  ];
 }
