+++ ui/app/routes/clients/client.js
@@ -9,16 +9,6 @@ export default class ClientRoute extends Route {
     return super.model(...arguments).catch(notifyError(this));
   }
 
-  breadcrumbs(model) {
-    if (!model) return [];
-    return [
-      {
-        label: model.get('shortId'),
-        args: ['clients.client', model.get('id')],
-      },
-    ];
-  }
-
   afterModel(model) {
     if (model && model.get('isPartial')) {
       return model.reload().then(node => node.get('allocations'));
