+++ ui/app/routes/csi/plugins.js
@@ -6,13 +6,6 @@ import notifyForbidden from 'nomad-ui/utils/notify-forbidden';
 export default class PluginsRoute extends Route.extend(WithForbiddenState) {
   @service store;
 
-  breadcrumbs = [
-    {
-      label: 'Storage',
-      args: ['csi.index'],
-    },
-  ];
-
   model() {
     return this.store.query('plugin', { type: 'csi' }).catch(notifyForbidden(this));
   }
