+++ ui/app/routes/jobs.js
@@ -1,12 +1,3 @@
 import Route from '@ember/routing/route';
-import classic from 'ember-classic-decorator';
 
-@classic
-export default class JobsRoute extends Route.extend() {
-  breadcrumbs = [
-    {
-      label: 'Jobs',
-      args: ['jobs.index'],
-    },
-  ];
-}
+export default class JobsRoute extends Route.extend() {}
