+++ ui/app/routes/jobs/job.js
@@ -2,7 +2,6 @@ import { inject as service } from '@ember/service';
 import Route from '@ember/routing/route';
 import RSVP from 'rsvp';
 import notifyError from 'nomad-ui/utils/notify-error';
-import { jobCrumbs } from 'nomad-ui/utils/breadcrumb-utils';
 import classic from 'ember-classic-decorator';
 
 @classic
@@ -10,8 +9,6 @@ export default class JobRoute extends Route {
   @service can;
   @service store;
   @service token;
-
-  breadcrumbs = jobCrumbs;
 
   serialize(model) {
     return { job_name: model.get('plainId') };
@@ -21,7 +18,11 @@ export default class JobRoute extends Route {
     const namespace = transition.to.queryParams.namespace || 'default';
     const name = params.job_name;
     const fullId = JSON.stringify([name, namespace]);
-
+    console.log(
+      this.store
+        .findRecord('job', fullId, { reload: true })
+        .then(job => console.log('model\n\n', job))
+    );
     return this.store
       .findRecord('job', fullId, { reload: true })
       .then(job => {
