+++ ui/app/routes/jobs/job/dispatch.js
@@ -4,13 +4,6 @@ import { inject as service } from '@ember/service';
 export default class DispatchRoute extends Route {
   @service can;
 
-  breadcrumbs = [
-    {
-      label: 'Dispatch',
-      args: ['jobs.job.dispatch'],
-    },
-  ];
-
   beforeModel() {
     const job = this.modelFor('jobs.job');
     const namespace = job.namespace.get('name');
