+++ ui/app/routes/optimize.js
@@ -9,13 +9,6 @@ import RSVP from 'rsvp';
 export default class OptimizeRoute extends Route {
   @service can;
 
-  breadcrumbs = [
-    {
-      label: 'Recommendations',
-      args: ['optimize'],
-    },
-  ];
-
   beforeModel() {
     if (this.can.cannot('accept recommendation')) {
       this.transitionTo('jobs');
