+++ ui/app/routes/optimize/summary.js
@@ -2,17 +2,6 @@ import Route from '@ember/routing/route';
 import notifyError from 'nomad-ui/utils/notify-error';
 
 export default class OptimizeSummaryRoute extends Route {
-  breadcrumbs(model) {
-    if (!model) return [];
-
-    return [
-      {
-        label: model.slug.replace('/', ' / '),
-        args: ['optimize.summary', model.slug],
-      },
-    ];
-  }
-
   async model({ jobNamespace, slug }) {
     const model = this.modelFor('optimize').summaries.find(
       summary => summary.slug === slug && summary.jobNamespace === jobNamespace
