+++ ui/app/routes/servers/server.js
@@ -1,14 +1,4 @@
 import Route from '@ember/routing/route';
 import WithModelErrorHandling from 'nomad-ui/mixins/with-model-error-handling';
 
-export default class ServerRoute extends Route.extend(WithModelErrorHandling) {
-  breadcrumbs(model) {
-    if (!model) return [];
-    return [
-      {
-        label: model.name,
-        args: ['servers.server', model.id],
-      },
-    ];
-  }
-}
+export default class ServerRoute extends Route.extend(WithModelErrorHandling) {}
