+++ ui/app/templates/allocations/allocation.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+{{outlet}}<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
