+++ ui/app/templates/allocations/allocation/task.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+{{outlet}}<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
