+++ ui/app/templates/components/app-breadcrumbs.hbs
@@ -1,13 +1,17 @@
-{{#each this.breadcrumbs as |breadcrumb index|}}
-  <li class="{{if (eq (inc index) this.breadcrumbs.length) "is-active"}}">
-    {{#if breadcrumb.isPending}}
-      <a href="#" aria-label="loading" data-test-breadcrumb="loading">&hellip;</a>
-    {{else}}
-      <LinkTo
-        @params={{breadcrumb.args}}
-        data-test-breadcrumb={{breadcrumb.args.firstObject}}>
-        {{breadcrumb.label}}
-      </LinkTo>
-    {{/if}}
-  </li>
-{{/each}}
+<Breadcrumbs as |bb|>
+  {{#each bb as |breadcrumb index|}}
+    {{#each breadcrumb.args.crumb as |b|}}
+      <li class="{{if (eq (inc index) this.b.length) "is-active"}}">
+        {{#if b.isPending}}
+          <a href="#" aria-label="loading" data-test-breadcrumb="loading">
+            …
+          </a>
+        {{else}}
+          <LinkTo @params={{b.args}} data-test-breadcrumb={{b.args.firstObject}}>
+            {{b.label}}
+          </LinkTo>
+        {{/if}}
+      </li>
+    {{/each}}
+  {{/each}}
+</Breadcrumbs>
