+++ ui/app/templates/csi/volumes.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+{{outlet}}<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
