+++ ui/app/templates/jobs/job.hbs
@@ -1,1 +1,1 @@
-{{outlet}}
+{{outlet}}<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
