+++ ui/app/templates/jobs/job/dispatch.hbs
@@ -1,5 +1,6 @@
 {{page-title "Dispatch new " this.model.name}}
 <JobSubnav @job={{this.model}} />
 <section class="section">
-  <JobDispatch @job={{this.model}}/>
+  <JobDispatch @job={{this.model}} />
 </section>
+<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
