+++ ui/app/templates/jobs/run.hbs
@@ -1,7 +1,5 @@
 {{page-title "Run a job"}}
 <section class="section">
-  <JobEditor
-    @job={{this.model}}
-    @context="new"
-    @onSubmit={{action this.onSubmit}} />
+  <JobEditor @job={{this.model}} @context="new" @onSubmit={{action this.onSubmit}} />
 </section>
+<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
