+++ ui/app/templates/servers.hbs
@@ -1,3 +1,4 @@
 <PageLayout>
   {{outlet}}
 </PageLayout>
+<Breadcrumb @bucket="navigation" @crumb={{this.breadcrumbs}} />
